Implement exception monitor
