+++ aiorospy/src/aiorospy/__init__.py
@@ -1,5 +1,6 @@
 from .action import AsyncActionClient, AsyncActionServer
 from .service import AsyncService, AsyncServiceProxy
 from .topic import AsyncSubscriber, AsyncPublisher
+from .helpers import ExceptionMonitor
 
 __all__ = ('AsyncSubscriber', 'AsyncService', 'AsyncServiceProxy', 'AsyncActionClient', 'AsyncActionServer')
